requirements/base.txt
Django==5.2.11
mssql-django==1.6
django-htmx==1.27.0
django-environ==0.12.0
waitress==3.0.2
openpyxl==3.1.5
